requirements.txt
jwcrypto==1.0
pycryptodome==3.11.0
requests==2.26.0
